+++ @vue-start/pro/src/comp/table/Table.tsx
@@ -58,6 +58,9 @@ export interface IOperateItem {
   sort?: number;
   per?: string; //权限字符串
   perSuffix?: string; //权限字符串后缀
+  //
+  tip?: string | VNode | ((record: Record<string, any>) => string | VNode); //tooltip提示
+  tipProps?: Record<string, any> | ((record: Record<string, any>) => Record<string, any>); //tooltip配置
 }
 
 /**
@@ -190,8 +193,10 @@ export const ProTable = defineComponent<ProTableProps>({
               disabled: isFunction(item.disabled) ? item.disabled(record) : item.disabled,
               loading: isFunction(item.loading) ? item.loading(record) : item.loading,
               extraProps: isFunction(item.extraProps) ? item.extraProps(record) : item.extraProps,
-              element: isFunction(item.element) ? () => item.element!(record, item) : item.element,
+              element: isFunction(item.element) ? (i) => item.element!(record, i) : item.element,
               onClick: () => handleOperateClick(record, item),
+              tip: isFunction(item.tip) ? item.tip(record) : item.tip,
+              tipProps: isFunction(item.tipProps) ? item.tipProps(record) : item.tipProps,
             } as IOpeItem;
           });