+++ src/views/element/operate/demo/extra.tsx
@@ -1,5 +1,5 @@
 /*---
-title: 展示、禁用
+title: 展示、禁用、tip
 ---*/
 import { defineComponent } from "vue";
 
@@ -7,7 +7,7 @@ export default defineComponent(() => {
   const items = [
     { label: "详情", value: "detail" },
     { label: "编辑", value: "edit", show: false },
-    { label: "删除", value: "delete", disabled: true },
+    { label: "删除", value: "delete", disabled: true, tip: "禁用删除" },
   ];
 
   return () => {