+++ src/views/element/table/demo/operate.tsx
@@ -1,6 +1,6 @@
 /*---
 title: 操作按钮
-desc: element方法可以重写dom；\nshow：boolean/()=>boolean，可以隐藏dom；
+desc: element方法可以重写dom；\nshow：boolean/()=>boolean，可以隐藏dom；\n提示配置tip
 ---*/
 import { defineComponent } from "vue";
 import { columns, dataSource } from "@/common/columns";
@@ -32,6 +32,12 @@ export default defineComponent(() => {
         value: "DELETE",
         label: "删除",
         disabled: (record: any) => record.id === "2", //第二条数据禁用删除
+        tip: (record: any) => {
+          if (record.id === "2") {
+            return "这条禁用删除";
+          }
+        },
+        // tipProps: { placement: "bottom" },
         onClick: (record: Record<string, any>) => {
           console.log("DELETE", record);
         },
@@ -41,13 +47,18 @@ export default defineComponent(() => {
         value: "custom",
         label: "自定义",
         //重写按钮
+        onClick: (record: Record<string, any>) => {
+          console.log("my custom", record);
+        },
         //id为1 禁用删除按钮
         element: (record: Record<string, any>, item: Record<string, any>) => {
           return (
             <span
               class={css({ color: "green", cursor: "pointer", marginLeft: 12 })}
               onClick={() => {
-                console.log(item.value, record);
+                //可以重写，也可以掉用原始方法
+                // console.log(item.value, record);
+                item.onClick?.();
               }}>
               {item.label}
             </span>